requirements.txt
streamlit
openai
streamlit-flow-component==1.5.0